requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiofiles==23.2.1
    # via gradio
altair==5.3.0
    # via gradio
annotated-types==0.7.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dnspython==2.6.1
    # via email-validator
email-validator==2.1.1
    # via fastapi
exceptiongroup==1.2.1
    # via anyio
fastapi==0.111.0
    # via gradio
fastapi-cli==0.0.4
    # via fastapi
ffmpy==0.3.2
    # via gradio
filelock==3.14.0
    # via huggingface-hub
fonttools==4.51.0
    # via matplotlib
fsspec==2024.5.0
    # via
    #   gradio-client
    #   huggingface-hub
gradio==4.31.4
    # via -r requirements.in
gradio-client==0.16.4
    # via gradio
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via
    #   fastapi
    #   gradio
    #   gradio-client
huggingface-hub==0.23.0
    # via
    #   gradio
    #   gradio-client
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
imageio==2.34.1
    # via
    #   -r requirements.in
    #   scikit-image
importlib-resources==6.4.0
    # via gradio
jinja2==3.1.4
    # via
    #   altair
    #   fastapi
    #   gradio
jsonschema==4.22.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
kiwisolver==1.4.5
    # via matplotlib
lazy-loader==0.4
    # via scikit-image
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   gradio
    #   jinja2
matplotlib==3.9.0
    # via gradio
mdurl==0.1.2
    # via markdown-it-py
networkx==3.3
    # via scikit-image
numpy==1.26.4
    # via
    #   altair
    #   contourpy
    #   gradio
    #   imageio
    #   matplotlib
    #   pandas
    #   scikit-image
    #   scipy
    #   tifffile
orjson==3.10.3
    # via
    #   fastapi
    #   gradio
packaging==24.0
    # via
    #   altair
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   lazy-loader
    #   matplotlib
    #   scikit-image
pandas==2.2.2
    # via
    #   altair
    #   gradio
pillow==10.3.0
    # via
    #   gradio
    #   imageio
    #   matplotlib
    #   scikit-image
pillow-avif-plugin==1.4.3
    # via -r requirements.in
pydantic==2.7.1
    # via
    #   fastapi
    #   gradio
pydantic-core==2.18.2
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.18.0
    # via rich
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.9
    # via
    #   fastapi
    #   gradio
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   gradio
    #   huggingface-hub
    #   uvicorn
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.1
    # via huggingface-hub
rich==13.7.1
    # via typer
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
ruff==0.4.4
    # via gradio
scikit-image==0.23.2
    # via -r requirements.in
scipy==1.13.0
    # via scikit-image
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.30
    # via -r requirements.in
starlette==0.37.2
    # via fastapi
tifffile==2024.5.10
    # via scikit-image
tomlkit==0.12.0
    # via gradio
toolz==0.12.1
    # via altair
tqdm==4.66.4
    # via huggingface-hub
typer==0.12.3
    # via
    #   fastapi-cli
    #   gradio
typing-extensions==4.11.0
    # via
    #   altair
    #   anyio
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   uvicorn
tzdata==2024.1
    # via pandas
ujson==5.10.0
    # via fastapi
urllib3==2.2.1
    # via
    #   gradio
    #   requests
uvicorn[standard]==0.29.0
    # via
    #   fastapi
    #   gradio
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==11.0.3
    # via
    #   gradio-client
    #   uvicorn
